+++ docs/changelog.rst
@@ -8,6 +8,10 @@ Changelog
 
 - New approach to stateful testing, based on the Hypothesis's ``RuleBasedStateMachine``. `#737`_
 
+**Changed**
+
+- The ``response_schema_conformance`` and ``content_type_conformance`` checks fail unconditionally if the input response has no ``Content-Type`` header. `#816`_
+
 **Deprecated**
 
 - Using ``stateful=Stateful.links`` in schema loaders and ``parametrize``. Use ``schema.as_state_machine().TestCase`` instead.
@@ -1411,6 +1415,7 @@ Deprecated
 .. _0.3.0: https://github.com/schemathesis/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/schemathesis/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#816: https://github.com/schemathesis/schemathesis/issues/816
 .. _#795: https://github.com/schemathesis/schemathesis/issues/795
 .. _#793: https://github.com/schemathesis/schemathesis/issues/793
 .. _#788: https://github.com/schemathesis/schemathesis/issues/788